Backend/requirements.txt
flask===3.0.0
charset-normalizer===3.4.0
click===8.1.7
colorama===0.4.6
docutils===0.16
idna===3.10
itsdangerous===2.2.0
jmespath===1.0.1
pyasn1===0.6.1
python-dateutil===2.9.0.post0
python-dotenv===1.0.1
requests===2.32.3
rsa===4.7.2
google-ai-generativelanguage===0.6.10
google-api-core===2.23.0
google-api-python-client===2.154.0
google-auth===2.36.0
google-auth-httplib2===0.2.0
google-generativeai===0.8.3
googleapis-common-protos===1.66.0
